+++ CHANGELOG.md
@@ -16,7 +16,21 @@ This project adheres to
     _Security_ in case of vulnerabilities.
  -->
 
-## [Unreleased](https://github.com/o1-labs/o1js/compare/08d8194c5...HEAD)
+## [Unreleased](https://github.com/o1-labs/o1js/compare/ff6c201b...HEAD)
+
+## [2.14.0](https://github.com/o1-labs/o1js/compare/c81f31ad0...ff6c201b) - 2026-03-16
+
+### Added
+
+- Collision-resistant version of Poseidon gadget for variable length inputs
+  `Poseidon.hashAnyLength()` (backwards incompatible with `Poseidon.hash()`) and
+  improved documentation with tests. https://github.com/o1-labs/o1js/pull/2753
+
+### Changed
+
+- Replace js-sha256 with @noble/hashes https://github.com/o1-labs/o1js/pull/2539
+
+## [2.13.1](https://github.com/o1-labs/o1js/compare/08d8194c5...c81f31ad0) - 2026-02-27
 
 ### Fixed